html-to-xlsx: Force date and datetime to be interpreted as UTC
The "moment" library will always interpret a date in the timezone of the operating system.  For linux, that defaults to UTC, but for Windows, that timezone can vary.  This change forces the cell's text to be interpreted as if the time is UTC so that the date is not offset by the operating system's timezone when it is written to the XLSX worksheet.