docx: fix case with template with multiple docxHtml calls and conditionals producing broken document
the problem was that one docxHtml call was in middle of loop.

the starting part which is the paragraph with __html_embed_container__="true" attribute was outside of loop, however the ending part <!--__html_embed_container__--> was part of a loop but also the closing paragraph, the loop ended in next paragraph.

to solve this we needed to normalize the closing comments when this happens, a handlebars helper helps us to keep track if the comments nodes are generated in a loop, if yes, only the latest item will generate the closing comment, so we dont end with multiple closing comments for just one paragraph, then we updated the postprocess part to assume that there can be multiple paragraphs between the starting part of paragraph and the closing comment.