fix sandbox tests with require of ESM modules on node 22
current node >=22.12 (LTS) supports using require to load ESM modules, however in our tests we were assuming it will always fail to load them.

to fix the tests we detect now if require module is supported, if yes we adapt the test based on the result.

this means on node >=22.12 the standard require can load es module, however the isolated require can not. adding support to load esm in the isolated require needs more effort, which does not seem worth it if we plan to land a major version soon. it looks better to wait for the next major version and have an updated isolated require that works with ESM but with no support for the older node.js versions that can not load ESM (v18.x.x)