docx: read exif orientation metadata of images for default display
fix #1230

some images (specially ones that comes from phone on portrait mode) differ on the way its data its stored vs the way they should be previewed. exif orientation on images contain the way the image buffer should be handled (rotated or flipped) in order for the image to display as expected.

we now read the exif orientation on images to tell the docx viewer to either rotate and/or flip images to finally achieve the correct default display of images.